+++ packages/expo-keep-awake/src/KeepAwake.types.ts
@@ -1,0 +1,10 @@
+// @needsAudit
+export type KeepAwakeEvent = {
+  /** Keep awake state. */
+  state: KeepAwakeEventState;
+};
+
+// @needsAudit
+export enum KeepAwakeEventState {
+  RELEASE = 'release',
+}