+++ packages/expo-keep-awake/src/index.ts
@@ -1,8 +1,20 @@
+import { Subscription, UnavailabilityError } from 'expo-modules-core';
 import { useEffect } from 'react';
 
 import ExpoKeepAwake from './ExpoKeepAwake';
+import { KeepAwakeEvent } from './KeepAwake.types';
 
 const ExpoKeepAwakeTag = 'ExpoKeepAwakeDefaultTag';
+
+/**
+ * Returns `true` on all platforms except [unsupported web browsers](https://caniuse.com/wake-lock).
+ */
+export async function isAvailableAsync(): Promise<boolean> {
+  if (ExpoKeepAwake.isAvailableAsync) {
+    return await ExpoKeepAwake.isAvailableAsync();
+  }
+  return true;
+}
 
 // @needsAudit
 /**
@@ -26,11 +38,13 @@ export function useKeepAwake(tag: string = ExpoKeepAwakeTag): void {
  * If the `tag` argument is specified, the screen will not sleep until you call `deactivateKeepAwake`
  * with the same `tag` argument. When using multiple `tags` for activation you'll have to deactivate
  * each one in order to re-enable screen sleep. If tag is unspecified, the default `tag` is used.
+ *
+ * Web support [is limited](https://caniuse.com/wake-lock).
+ *
  * @param tag *Optional* - Tag to lock screen sleep prevention. If not provided, the default tag is used.
  */
-
-export function activateKeepAwake(tag: string = ExpoKeepAwakeTag): void {
-  if (ExpoKeepAwake.activate) ExpoKeepAwake.activate(tag);
+export async function activateKeepAwake(tag: string = ExpoKeepAwakeTag): Promise<void> {
+  if (ExpoKeepAwake.activate) await ExpoKeepAwake.activate(tag);
 }
 
 // @needsAudit
@@ -43,3 +57,31 @@ export function activateKeepAwake(tag: string = ExpoKeepAwakeTag): void {
 export function deactivateKeepAwake(tag: string = ExpoKeepAwakeTag): void {
   if (ExpoKeepAwake.deactivate) ExpoKeepAwake.deactivate(tag);
 }
+
+/**
+ * Observe changes to the keep awake timer.
+ * On web, this changes when navigating away from the active window/tab. No-op on native.
+ *
+ * @example
+ * ```ts
+ * KeepAwake.addListener(({ state }) => {
+ *   // ...
+ * });
+ * ```
+ */
+export function addListener(
+  tagOrListener: string | ((event: KeepAwakeEvent) => void),
+  listener?: (event: KeepAwakeEvent) => void
+): Subscription {
+  // Assert so the type is non-nullable.
+  if (!ExpoKeepAwake.addListener) {
+    throw new UnavailabilityError('ExpoKeepAwake', 'addListener');
+  }
+
+  const tag = typeof tagOrListener === 'string' ? tagOrListener : ExpoKeepAwakeTag;
+  const _listener = typeof tagOrListener === 'function' ? tagOrListener : listener;
+
+  return ExpoKeepAwake.addListener(tag, 'ExpoNavigationBar.didChange', _listener);
+}
+
+export * from './KeepAwake.types';