+++ packages/expo-keep-awake/CHANGELOG.md
@@ -4,8 +4,13 @@
 
 ### 🛠 Breaking changes
 
+- `KeepAwake.activateKeepAwake` now returns a promise to accommodate web.
+
 ### 🎉 New features
 
+- Added web support.
+- Added `KeepAwake.isAvailableAsync` which returns false on certain web browsers.
+- Added `KeepAwake.addListener` to observe state changes on web.
 - Native module on iOS is now written in Swift using the new API. ([#15705](https://github.com/expo/expo/pull/15705) by [@tsapeta](https://github.com/tsapeta))
 
 ### 🐛 Bug fixes